% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_xgboost.R
\name{pred_xgboost_indiv}
\alias{pred_xgboost_indiv}
\title{Classification from individually specified features}
\usage{
pred_xgboost_indiv(Sepal.Length = 0, Sepal.Width = 0,
  Petal.Length = 0, Petal.Width = 0)
}
\description{
Classification from individually specified features
}
