% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_base.R
\name{preprocessing}
\alias{preprocessing}
\title{Preprocessing de base
Ce programme a pour but de faire le preprocessing commun au 2 modeles}
\usage{
preprocessing(data, train_mode = TRUE, list_objects = NULL)
}
\description{
Preprocessing de base
Ce programme a pour but de faire le preprocessing commun au 2 modeles
}
