% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_classif.R
\name{preprocessing_classif}
\alias{preprocessing_classif}
\title{Preprocessing classification
Ce programme fait le preprocessing pour le modele de classification seulement
On assume ici que le preprocessing de base a été roulé}
\usage{
preprocessing_classif(data, train_mode = TRUE, list_objects = NULL)
}
\description{
Preprocessing classification
Ce programme fait le preprocessing pour le modele de classification seulement
On assume ici que le preprocessing de base a été roulé
}
