% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_regression.R
\name{preprocessing_regression}
\alias{preprocessing_regression}
\title{Preprocessing regression
Ce programme fait le preprocessing pour le modele de regression seulement
On assume ici que le preprocessing de base a été roulé}
\usage{
preprocessing_regression(data, train_mode = TRUE, list_objects = NULL)
}
\description{
Preprocessing regression
Ce programme fait le preprocessing pour le modele de regression seulement
On assume ici que le preprocessing de base a été roulé
}
