% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_bixi.R
\name{bixikwargs}
\alias{bixikwargs}
\title{Classification from individually specified features}
\usage{
bixikwargs(start_date, start_station_code, end_date, end_station_code,
  duration_sec, is_member)
}
\description{
Classification from individually specified features
}
